[autogen][cpp] Allow setters chaining and introduce motor commands
This is a two part change. The first makes attribute setters return
reference to the device. This makes possible chaining of setters:

    d = large_motor();
    d.set_speed_regulation_enabled("on").set_speed_sp(600);

The second part introduces autogenerated methods corresponding to motor
commands. The methods take no arguments and just set the command
attribute to the appropriate value. This may be comnined with a chain of
attribute setters:

    d.set_time_sp(1000).set_duty_cycle_sp(75).run_timed();

I had to change the name of "float" command for servo_motor class to
"stop" since "float" is a keyword in C++.