[cpp] Solve the problem with led::flash
The problem was reported in ddemidov/ev3dev-lang-python#7.
Each time a led trigger is set to `timer`, the driver recreates
`delay_on` and `delay_off` attributes, which invalidates the
cached file handles.

When such stale file handle is used, an exception is thrown.

This commit modifies device::{get,set}_attr_int methods so that
an attempt to reopen the file is made in such case.